folder structure change for models and useEffect fixes
