+++ litmus-portal/frontend/src/components/Sections/ChaosWorkflows/BrowseWorkflow/TableData.tsx
@@ -7,7 +7,10 @@ import {
 } from '@material-ui/core';
 import MoreVertIcon from '@material-ui/icons/MoreVert';
 import React from 'react';
-import { ExecutionData, WorkflowRun } from '../../../../models/workflowData';
+import {
+  ExecutionData,
+  WorkflowRun,
+} from '../../../../models/graphql/workflowData';
 import { history } from '../../../../redux/configureStore';
 import timeDifferenceForDate from '../../../../utils/datesModifier';
 import LinearProgressBar from '../../ReturningHome/ProgressBar/LinearProgressBar';
