+++ litmus-portal/frontend/src/components/Sections/Settings/AccountsTab/PersonalDetails/index.tsx
@@ -1,17 +1,16 @@
+import { useQuery } from '@apollo/client';
 import { Button, Typography } from '@material-ui/core';
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { useQuery } from '@apollo/client';
+import Unimodal from '../../../../../containers/layouts/Unimodal';
+import { GET_USER } from '../../../../../graphql';
 import {
-  UserData,
-  CurrentUserDetails,
   CurrentUserDedtailsVars,
-} from '../../../../../models/user';
+  CurrentUserDetails,
+} from '../../../../../models/graphql/user';
 import { RootState } from '../../../../../redux/reducers';
 import UserDetails from '../../UserManagementTab/CreateUser/UserDetails';
 import useStyles from './styles';
-import Unimodal from '../../../../../containers/layouts/Unimodal';
-import { GET_USER } from '../../../../../graphql';
 
 interface personaData {
   email: string;
@@ -22,8 +21,7 @@ interface personaData {
 // Displays the personals details on the "accounts" tab
 const PersonalDetails: React.FC = () => {
   const classes = useStyles();
-  const userData: UserData = useSelector((state: RootState) => state.userData);
-  const { username } = userData;
+  const username = useSelector((state: RootState) => state.userData.username);
 
   // Query to get user details
   const { data } = useQuery<CurrentUserDetails, CurrentUserDedtailsVars>(
