+++ litmus-portal/frontend/src/models/redux/analytics.ts
