+++ litmus-portal/frontend/src/models/redux/user.ts
@@ -1,5 +1,3 @@
-import { Project } from './project';
-
 export interface UserData {
   selectedProjectID: string;
   username: string;
@@ -7,28 +5,6 @@ export interface UserData {
   token: string;
   selectedProjectName: string;
   userRole: string;
-}
-
-export interface UserDetails {
-  username: string;
-  projects: Project[];
-  name: string;
-  email: string;
-  company_name: string;
-  updated_at: string;
-  created_at: string;
-  removed_at: string;
-  is_email_verified: string;
-  state: string;
-  role: string;
-}
-
-export interface CurrentUserDetails {
-  getUser: UserDetails;
-}
-
-export interface CurrentUserDedtailsVars {
-  username: string;
 }
 
 export enum UserActions {
