+++ litmus-portal/frontend/src/redux/reducers/index.ts
@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux';
-import { CommunityData } from '../../models/analytics';
-import { UserData } from '../../models/user';
-import { WorkflowData } from '../../models/workflow';
-import { Node } from '../../models/workflowData';
+import { Node } from '../../models/graphql/workflowData';
+import { CommunityData } from '../../models/redux/analytics';
+import { UserData } from '../../models/redux/user';
+import { WorkflowData } from '../../models/redux/workflow';
 import * as analyticsReducer from './analytics';
 import * as nodeSelectionReducer from './nodeSelection';
 import * as userReducer from './user';
